re-organise file structure a little to be clearer
