+++ src/pages/company.js
