+++ src/pages/contact.js
@@ -1,53 +1,9 @@
-/* globals interactions: true */
 require('babel-polyfill');
 const React = require('react');
 const ReactDOM = require('react-dom');
-const { addClass, removeClass } = require('../lib/elementstuff');
-const InteractionTable = require('../components/interactiontable.component.js');
-const Tabs = require('../controls/tabs');
+const Routes = require('./../reactrouting/contactroutes').Routes;
 
-const archiveForm = document.getElementById('archive-details');
-const archiveButton = document.getElementById('archive-reveal-button');
-const cancelButton = document.getElementById('cancel-archive-button');
-const archiveReasonElement = document.getElementById('archived_reason');
-const archiveReasonGroup = document.getElementById('archived_reason-wrapper');
-
-new Tabs(document.querySelector('.js-tabs'));
-
-if (interactions && interactions.length > 0) {
-  ReactDOM.render(
-    <InteractionTable interactions={interactions} />,
-    document.querySelector('#interaction-table-wrapper')
-  );
-}
-
-function revealArchive(event) {
-  event.preventDefault();
-  removeClass(archiveForm, 'hidden');
-  addClass(archiveButton, 'hidden');
-  archiveForm.scrollIntoView();
-}
-
-function hideArchive(event) {
-  event.preventDefault();
-  addClass(archiveForm, 'hidden');
-  removeClass(archiveButton, 'hidden');
-}
-
-function showArchiveError() {
-  addClass(archiveReasonGroup, 'error');
-}
-
-function submitArchiveForm(event) {
-  let reason = archiveReasonElement.options[archiveReasonElement.selectedIndex].value;
-  if (!reason) {
-    event.preventDefault();
-    showArchiveError();
-  }
-}
-
-if (archiveButton) {
-  archiveButton.addEventListener('click', revealArchive);
-  cancelButton.addEventListener('click', hideArchive);
-  archiveForm.addEventListener('submit', submitArchiveForm);
-}
+ReactDOM.render(
+  <Routes />,
+  document.getElementById('contact'),
+);
