--- src/pages/contact.page.js
@@ -1,9 +1,0 @@
-require('babel-polyfill');
-const React = require('react');
-const ReactDOM = require('react-dom');
-const Routes = require('./../reactrouting/contactroutes').Routes;
-
-ReactDOM.render(
-  <Routes />,
-  document.getElementById('contact'),
-);
